+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Changes
 
 - **Merged some changes from Woof post-11.1.4**
+- **Changed mouselook range to [-100, 100]**
 - **_'SUMMON'_ now reports the last summoned mobj's type**
 - **_'SUMMONE'_ and _'SUMMONF'_ now print a message requesting the mobj index**
 - **Mobjs summoned with _'SUMMON'_ now inherit the player's angle**